requirements.txt

streamlit
pillow
pytesseract
